test/unit/modules/monsters/assets/destroy_monster_asset.spec.ts
import { DestroyMonsterAsset } from '../../../../../src/app/modules/monsters/assets/destroy_monster_asset';

describe('DestroyMonsterAsset', () => {
	let transactionAsset: DestroyMonsterAsset;

	beforeEach(() => {
		transactionAsset = new DestroyMonsterAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(1);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('destroyMonster');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
			it.todo('should throw errors for invalid schema');
			it.todo('should be ok for valid schema');
		});
	});

	describe('apply', () => {
		describe('valid cases', () => {
			it.todo('should update the state store');
		});

		describe('invalid cases', () => {
			it.todo('should throw error');
		});
	});
});
